+++ lib/variable/operations_common.h
@@ -11,6 +11,10 @@
 
 namespace scipp::variable {
 
+/// Sum elements of `var` and add to `summed` along dims
+/// present in `var` but not in `summed`.
+SCIPP_VARIABLE_EXPORT void sum_into(Variable &summed, const Variable &var);
+
 // Helpers for in-place reductions and reductions with groupby.
 SCIPP_VARIABLE_EXPORT Variable mean_impl(const Variable &var, const Dim dim,
                                          const Variable &masks_sum);